refactor(validator): 유효성 검증 로직을 전담하는 PurchaseAmountValidator 클래스로 분리
